working on tinyworlds config
